Rename viewer surface namespace to graphistry_frontend
